Add user identity to Comments Logic.

+++ resources/comment.py
@@ -2,7 +2,8 @@ from flask import Response, request
 from database.models import Comment, Post
 from flask_restful import Resource
 from mongoengine.queryset.visitor import Q
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_identity
+from utils.user import get_user_id, is_comment_author
 
 class CommentsApi(Resource):
     decorators = [jwt_required()]
@@ -18,6 +19,7 @@ class CommentsApi(Resource):
         comment = Comment(**body)
         post = Post.objects.get(id=id)
         comment.post = post
+        comment.user = get_user_id(get_jwt_identity())
         comment.save()
         comment_id = comment.id
         return {'id': str(comment_id), 'post_id': id}, 201
@@ -27,9 +29,15 @@ class CommentApi(Resource):
     decorators = [jwt_required()]
 
     def delete(self, id, comment_id):
-        comment = Comment.objects(Q(id=comment_id)| Q(post=id))
-        if comment:
-            comment.delete()
-            return '', 200
-        else:
-            return '', 404
+        try:
+            comment = Comment.objects(Q(id=comment_id)| Q(post=id))
+            if comment:
+                if is_comment_author(get_jwt_identity(), comment_id):
+                    comment.delete()
+                    return {'msg': 'Comment deleted'}, 200
+                else:
+                    return {'error': 'Access denied'}, 403
+            else:
+                return {'error': 'Comment Not found'}, 404
+        except:
+            return {'error': 'Unknown error'}, 404
